fresh/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.5
django-cors-headers==3.3.0
django-mysql==3.6.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
idna==2.9
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysqlclient @ file:///C:/Users/Santanu/project/fresh/mysqlclient-1.4.6-cp38-cp38-win32.whl
openapi-codec==1.3.2
Pillow==7.1.2
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2019.3
requests==2.23.0
simplejson==3.17.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
